+++ packages/Manager/src/test/kotlin/nz/govt/eop/consumers/observations/ObservationsConsumerTest.kt
@@ -1,4 +1,4 @@
-package nz.govt.eop.consumers
+package nz.govt.eop.consumers.observations
 
 import io.kotest.matchers.shouldBe
 import java.math.BigDecimal
